hshl/ti/11/requirements.txt
psycopg2-binary
folium
pandas
seaborn
matplotlib
sqlalchemy
